fix int64 casts with NTL

+++ linbox/ring/ntl/ntl-gf2e.h
@@ -213,7 +213,7 @@ public :
                 if(_seed == 0)
                     NTL::SetSeed(NTL::to_ZZ(time(0)));
                 else
-                    NTL::SetSeed(NTL::to_ZZ(_seed));
+                    NTL::SetSeed(NTL::to_ZZ(static_cast<long unsigned int>(_seed)));
             }
 
 		UnparametricRandIter<NTL::GF2E>(const UnparametricRandIter<NTL::GF2E>& R) :
@@ -223,7 +223,7 @@ public :
                 if(_seed == 0)
                     NTL::SetSeed(NTL::to_ZZ(time(0)));
                 else
-                    NTL::SetSeed(NTL::to_ZZ(_seed));
+                    NTL::SetSeed(NTL::to_ZZ(static_cast<long unsigned int>(_seed)));
             }
 
 		Element& random (Element& x) const
